requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.3
protobuf==3.10.0
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
pandas=0.20.3
Keras==2.1.2
scikit-learn==0.22.1